requirements.txt
boto3
web3
python-dotenv